requirements.txt
#

# Spotify client
spotipy==2.18.0

# Set up variables as environment variables
python-dotenv==0.18.0